+++ src/main/kotlin/com/github/pooryam92/vimcoach/services/TipLoaderService.kt
@@ -1,6 +1,6 @@
 package com.github.pooryam92.vimcoach.services
 
 interface TipLoaderService {
-    fun loadTips()
-    fun refetchTips()
+    fun refetchTips(): TipLoadResult
+    fun checkForUpdates(): TipLoadResult
 }